add CompareBuySimilar.lua for logic related to buying similar items
